Renaming and cleaning up.
